Move formatBlock to its own file
