+++ programs/compressor/CMakeLists.txt
@@ -5,7 +5,6 @@ set (CLICKHOUSE_COMPRESSOR_SOURCES Compressor.cpp)
 set (CLICKHOUSE_COMPRESSOR_LINK
     PRIVATE
         boost::program_options
-        clickhouse_dictionaries
         clickhouse_parsers
         dbms
 )
