+++ programs/git-import/CMakeLists.txt
@@ -3,7 +3,6 @@ set (CLICKHOUSE_GIT_IMPORT_SOURCES git-import.cpp)
 set (CLICKHOUSE_GIT_IMPORT_LINK
     PRIVATE
         boost::program_options
-        clickhouse_dictionaries
         dbms
 )
