+++ programs/keeper-converter/CMakeLists.txt
@@ -3,7 +3,6 @@ set (CLICKHOUSE_KEEPER_CONVERTER_SOURCES KeeperConverter.cpp)
 set (CLICKHOUSE_KEEPER_CONVERTER_LINK
     PRIVATE
         boost::program_options
-        clickhouse_dictionaries
         dbms
 )
