+++ programs/library-bridge/CMakeLists.txt
@@ -15,7 +15,6 @@ endif ()
 add_executable(clickhouse-library-bridge ${CLICKHOUSE_LIBRARY_BRIDGE_SOURCES})
 
 target_link_libraries(clickhouse-library-bridge PRIVATE
-    clickhouse_dictionaries
     daemon
     dbms
     bridge
