+++ programs/local/CMakeLists.txt
@@ -6,7 +6,6 @@ set (CLICKHOUSE_LOCAL_LINK
         clickhouse_aggregate_functions
         clickhouse_common_config
         clickhouse_common_io
-        clickhouse_dictionaries
         clickhouse_functions
         clickhouse_parsers
         clickhouse_storages_system
