+++ programs/obfuscator/CMakeLists.txt
@@ -3,7 +3,6 @@ set (CLICKHOUSE_OBFUSCATOR_SOURCES Obfuscator.cpp)
 set (CLICKHOUSE_OBFUSCATOR_LINK
     PRIVATE
         boost::program_options
-        clickhouse_dictionaries
         dbms
 )
