+++ src/DataStreams/formatBlock.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <DataStreams/IBlockStream_fwd.h>
+
+namespace DB
+{
+void formatBlock(BlockOutputStreamPtr & out, const Block & block);
+
+}
