+++ src/Dictionaries/DictionarySourceHelpers.cpp
@@ -1,7 +1,7 @@
 #include "DictionarySourceHelpers.h"
 #include <Columns/ColumnsNumber.h>
 #include <Core/ColumnWithTypeAndName.h>
-#include <DataStreams/IBlockOutputStream.h>
+#include <DataStreams/IBlockStream_fwd.h>
 #include <DataTypes/DataTypesNumber.h>
 #include <IO/WriteHelpers.h>
 #include "DictionaryStructure.h"
@@ -16,14 +16,6 @@ namespace DB
 namespace ErrorCodes
 {
     extern const int SIZES_OF_COLUMNS_DOESNT_MATCH;
-}
-
-void formatBlock(BlockOutputStreamPtr & out, const Block & block)
-{
-    out->writePrefix();
-    out->write(block);
-    out->writeSuffix();
-    out->flush();
 }
 
 /// For simple key
