+++ src/Dictionaries/HTTPDictionarySource.cpp
@@ -1,6 +1,7 @@
 #include "HTTPDictionarySource.h"
 #include <DataStreams/IBlockOutputStream.h>
 #include <DataStreams/OwningBlockInputStream.h>
+#include <DataStreams/formatBlock.h>
 #include <IO/ConnectionTimeouts.h>
 #include <IO/ConnectionTimeoutsContext.h>
 #include <IO/ReadWriteBufferFromHTTP.h>
